+++ example/bridge_client_example_1.dart
@@ -6,12 +6,12 @@ void main() {
     transaction: BridgeClientTransaction(
       logInterface: const ConsoleLogInterface(),
       topic: 'CimZzz',
-      remoteTopic: 'CimZzz',
-      transportAddress: 'virtual-lightning.com',
-      transportPort: 80,
+      remoteTopic: 'TiwZzz',
+      peerAddress: 'virtual-lightning.com',
+      peerPort: 80,
+      isPeerCustomTransport: true,
       bridgeAddress: '127.0.0.1',
       bridgePort: 9000,
-      isCustomTransport: true,
       rsaPublicKeyPath: '/Users/wangyanxiong/Documents/IDEAProject/transport/rsa_public_key.pem',
     )
   ).startServer();