+++ example/bridge_client_example_2.dart
@@ -7,8 +7,6 @@ void main() {
       logInterface: const ConsoleLogInterface(),
       topic: 'TiwZzz',
       remoteTopic: 'CimZzz',
-      transportAddress: 'virtual-lightning.com',
-      transportPort: 80,
       bridgeAddress: '127.0.0.1',
       bridgePort: 9000,
       rsaPublicKeyPath: '/Users/wangyanxiong/Documents/IDEAProject/transport/rsa_public_key.pem',